Move examples definition to allow feature flagging
