+++ client/branded/src/search-ui/input/experimental/codemirror/example-suggestions.ts
@@ -1,9 +1,7 @@
 import type { Extension } from '@codemirror/state'
 import { sampleSize } from 'lodash'
 
-import { FILTERS, FilterType } from '@sourcegraph/shared/src/search/query/filters'
 import type { Token } from '@sourcegraph/shared/src/search/query/token'
-import { resolveFilterMemoized } from '@sourcegraph/shared/src/search/query/utils'
 
 import { getQueryInformation } from '../../codemirror/parsedQuery'
 import { queryRenderer } from '../optionRenderer'
@@ -15,7 +13,7 @@ import {
     selectionListener,
 } from '../suggestionsExtension'
 
-interface Example {
+export interface Example {
     label: string
     description: string
     snippet?: string
@@ -28,64 +26,18 @@ interface Example {
 const REFRESH_TIMEOUT = 60000 // 1 minute
 const EXAMPLES_TO_SHOW = 3
 
-const examples: Example[] = [
-    {
-        label: 'file:has.owner()',
-        snippet: 'file:has.owner(${}) ${}',
-        description: 'Search code ownership',
-        valid: tokens => !tokens.some(token => token.type === 'filter' && token.value?.value.startsWith('has.owner(')),
-    },
-    {
-        label: 'repo:has.path()',
-        snippet: 'repo:has.path(${}) ${}',
-        description: 'Search in repositories containing a path',
-        valid: tokens => !tokens.some(token => token.type === 'filter' && token.value?.value.startsWith('has.path(')),
-    },
-    {
-        label: 'repo:has.content()',
-        snippet: 'repo:has.content(${}) ${}',
-        description: 'Search in repositories with files having specific contents',
-        valid: tokens =>
-            !tokens.some(token => token.type === 'filter' && token.value?.value.startsWith('has.content(')),
-    },
-    {
-        label: '-file:',
-        description: FILTERS[FilterType.file].description(true),
-        valid: tokens => !tokens.some(token => token.type === 'filter' && token.field.value === '-file'),
-    },
-    {
-        label: '-repo:',
-        description: FILTERS[FilterType.repo].description(true),
-        valid: tokens => !tokens.some(token => token.type === 'filter' && token.field.value === '-repo'),
-    },
-    {
-        label: 'repo:my-org.*/.*-cli$',
-        // eslint-disable-next-line no-template-curly-in-string
-        snippet: 'repo:${my-org.*/.*-cli$} ${}',
-        description: 'Search in repositories matching a pattern',
-        valid: tokens =>
-            !tokens.some(
-                token => token.type === 'filter' && resolveFilterMemoized(token.field.value)?.type === FilterType.repo
-            ),
-    },
-    {
-        label: 'type:diff select:commit.diff.removed TODO',
-        // eslint-disable-next-line no-template-curly-in-string
-        snippet: 'type:diff select:commit.diff.removed repo:${my-repo} TODO ${}',
-        description: 'Find commits that removed "TODO"',
-        valid: tokens => !tokens.some(token => token.type === 'filter' && token.value?.value.startsWith('commit.diff')),
-    },
-]
-
-const choosenExamples = {
-    examples: sampleSize(examples, EXAMPLES_TO_SHOW),
-    timestamp: Date.now(),
-}
-
 export function exampleSuggestions(config: {
     getUsedExamples: () => Set<string>
     markExampleUsed: (example: string) => void
+    examples: Example[]
 }): Extension {
+    const { getUsedExamples, markExampleUsed, examples } = config
+
+    const choosenExamples = {
+        examples: sampleSize(examples, EXAMPLES_TO_SHOW),
+        timestamp: Date.now(),
+    }
+
     const hideAction: Action = {
         type: 'command',
         name: 'Hide',
@@ -106,7 +58,7 @@ export function exampleSuggestions(config: {
                     return null
                 }
 
-                const usedExamples = config.getUsedExamples()
+                const usedExamples = getUsedExamples()
                 const validExamples = examples.filter(
                     example => !usedExamples.has(example.label) && example.valid(tokens)
                 )
@@ -153,7 +105,7 @@ export function exampleSuggestions(config: {
         }),
         selectionListener.of(({ option }) => {
             if (option.kind === 'example') {
-                config.markExampleUsed(option.label)
+                markExampleUsed(option.label)
             }
         }),
     ]
